+++ NEWS.md
@@ -6,6 +6,7 @@
 * Removed the chi-squared test from `plot_mosaic()`.
 * The `correct` argument of `compare_sequences()` has been renamed to `adjust` to align with other similar arguments.
 * Added several new arguments to the `plot` method of `tna` objects for plotting heterogeneous networks.
+* The `plot` method for `communities` now uses the first available method by default.
 
 # tna 1.1.0